Try to fix install_batik once again

+++ .cloud66/install_batik.sh
@@ -1,6 +1,8 @@
-apt-get install libbatik-java -y
+if [ -d "/usr/local/bin/batik" ]; then
+  apt-get install libbatik-java -y
 
-wget http://www-us.apache.org/dist/xmlgraphics/batik/binaries/batik-bin-1.8.tar.gz
-tar xzvf batik-bin-1.8.tar.gz
-rm batik-bin-1.8.tar.gz
-mv batik-1.8/ /usr/local/bin/batik
+  wget http://www-us.apache.org/dist/xmlgraphics/batik/binaries/batik-bin-1.8.tar.gz
+  tar xzvf batik-bin-1.8.tar.gz
+  rm batik-bin-1.8.tar.gz
+  mv batik-1.8/ /usr/local/bin/batik
+fi
